services/ai/requirements.txt
# AI Service Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0

# Firebase
firebase-admin==6.2.0

# AI - OpenRouter (via aiohttp)
aiohttp==3.9.1

# Image processing
Pillow==10.1.0
python-multipart==0.0.6

# Redis caching
redis==5.0.1

# HTTP client
httpx==0.25.2

# Document processing
PyPDF2==3.0.1
